changed save track screen, using header from react-native-navigation instead custom header
