+++ messages/en.json
@@ -44,7 +44,7 @@
     "message": "Start Tracks"
   },
   "Modal.GPSEnable.button.loading": {
-    "message": "Loading..."
+    "message": "Loading…"
   },
   "Modal.GPSEnable.button.stop": {
     "message": "Stop Tracks"
