+++ src/frontend/Navigation/ScreenGroups/AppScreens.tsx
@@ -367,13 +367,7 @@ export const createDefaultScreenGroup = (
       component={GpsModal}
       options={createGpsModalNavigationOptions({intl})}
     />
-    <RootStack.Screen
-      name="SaveTrack"
-      component={SaveTrackScreen}
-      options={{
-        headerShown: false,
-      }}
-    />
+    <RootStack.Screen name="SaveTrack" component={SaveTrackScreen} />
     <RootStack.Screen
       name="InviteDeclined"
       component={InviteDeclined}
